requirements.txt
feedparser>=6.0.10
requests>=2.31.0
jinja2>=3.1.2
python-dateutil>=2.8.2
openai>=1.0.0
